+++ languagetool-server/src/main/java/org/languagetool/server/Server.java
@@ -93,7 +93,7 @@ abstract class Server {
   protected RequestLimiter getRequestLimiterOrNull(HTTPServerConfig config) {
     int requestLimit = config.getRequestLimit();
     int requestLimitPeriodInSeconds = config.getRequestLimitPeriodInSeconds();
-    if (requestLimit > 0 || requestLimitPeriodInSeconds > 0) {
+    if (requestLimit > 0 && requestLimitPeriodInSeconds > 0) {
       return new RequestLimiter(requestLimit, requestLimitPeriodInSeconds);
     }
     return null;
@@ -101,10 +101,10 @@ abstract class Server {
 
   @Nullable
   protected ErrorRequestLimiter getErrorRequestLimiterOrNull(HTTPServerConfig config) {
-    int requestLimit = config.getRequestLimit();
+    int requestLimit = config.getTimeoutRequestLimit();
     int requestLimitPeriodInSeconds = config.getRequestLimitPeriodInSeconds();
-    if (requestLimit > 0 || requestLimitPeriodInSeconds > 0) {
-      return new ErrorRequestLimiter(2, 60);  // TODO: use config values
+    if (requestLimit > 0 && requestLimitPeriodInSeconds > 0) {
+      return new ErrorRequestLimiter(requestLimit, requestLimitPeriodInSeconds);
     }
     return null;
   }
@@ -123,7 +123,8 @@ abstract class Server {
     System.out.println("                 'maxCheckThreads' - maximum number of threads working in parallel (optional)");
     System.out.println("                 'cacheSize' - size of internal cache in number of sentences (optional, default: 0)");
     System.out.println("                 'requestLimit' - maximum number of requests (optional)");
-    System.out.println("                 'requestLimitPeriodInSeconds' - time period to which requestLimit applies (optional)");
+    System.out.println("                 'timeoutRequestLimit' - maximum number of timeout request (optional)");
+    System.out.println("                 'requestLimitPeriodInSeconds' - time period to which requestLimit and timeoutRequestLimit applies (optional)");
     System.out.println("                 'languageModel' - a directory with '1grams', '2grams', '3grams' sub directories which contain a Lucene index");
     System.out.println("                  each with ngram occurrence counts; activates the confusion rule if supported (optional)");
     System.out.println("                 'maxWorkQueueSize' - reject request if request queue gets larger than this (optional)");